chore(build): Statically build libbpf-sys
This avoids the requirement for users to have both libelf and zlib present on their system when running the binary. I checked and the resulting binary size increases by about 1MB for static linking vs dynamic linking. This is acceptable I think.
Build times suffer slightly, but this is Rust - choose Go if you want a fast compiler :)